refactor: rename lua ai runtime modules
