+++ moongate_data/scripts/ai/brains/ai_melee.lua
@@ -1,6 +1,6 @@
-local fsm = require("ai.modernuo.fsm")
-local movement = require("ai.modernuo.movement")
-local targeting = require("ai.modernuo.targeting")
+local fsm = require("ai.runtime.fsm")
+local movement = require("ai.runtime.movement")
+local targeting = require("ai.runtime.targeting")
 
 ai_melee = {}
