+++ moongate_data/scripts/ai/init.lua
@@ -1,8 +1,8 @@
 require("ai.behavior")
 require("ai.behaviors")
-require("ai.modernuo.fsm")
-require("ai.modernuo.targeting")
-require("ai.modernuo.movement")
+require("ai.runtime.fsm")
+require("ai.runtime.targeting")
+require("ai.runtime.movement")
 require("ai.brains.ai_melee")
 require("ai.brains.ai_archer")
 require("ai.brains.ai_animal")
